requirements.txt
requests>=2.32.0
beautifulsoup4>=4.12.3
trafilatura>=1.9.0
python-dotenv>=1.0.1
reportlab>=4.2.5